codeforces/cpp/1800-1899/1850G.cpp

+++ codeforces/cpp/1800-1899/1850G.cpp
@@ -5,14 +5,14 @@
 int main() {
     int t;
     int n;
-    int x, y;
+    long long x, y;
     std::cin >> t;
 
     for (int i = 0; i < t; i++) {
         std::cin >> n;
 
-        std::unordered_map<int, int> same_x, same_y, positive_45, negative_45;
-        int response = 0;
+        std::unordered_map<long long, long long> same_x, same_y, positive_45, negative_45;
+        long long response = 0;
 
         for (int j = 0; j < n; j++) {
             std::cin >> x >> y;
